Improve output video component
